+++ src/components/NQList.vue
@@ -15,7 +15,7 @@
       <div class="col-12">
         <div v-masonry transition-duration="0.3s" item-selection=".media-cardl">
           <q-card v-masonry-tile v-for="item in showItems" :key="item._id" color="primary" v-bind:class="[item.type] + 'l'" class="media-cardl" @click.native="openItem(item.media, item.type)">
-            <!-- <q-icon name="fas fa-plus" color="positive" class="float-right cursor-pointer" style="margin-top: 5px; margin-right: 5px;" /> -->
+            <q-btn v-if="addButton === ''" icon="fas fa-plus" color="positive" class="float-right cursor-pointer" style="margin-bottom: 5px; margin-left: 5px;" dense size="sm" />
             <q-card-media v-if="item.type == 'book' || item.type == 'movie' || item.type == 'video' || item.type == 'article' || item.type == 'image'">
               <img :src="item.media.thumbURL" />
             </q-card-media>
@@ -24,8 +24,18 @@
               <span v-if="item.type == 'book' || item.type == 'movie' || item.type == 'video' || item.type == 'article' || item.type == 'discourse'" v-for="author in item.media.author" :key="author" slot="subtitle">{{ author }}</span>
               <span v-if="item.type == 'note'" slot="subtitle">{{ item.media.text }}</span>
             </q-card-title>
-            <q-card-main v-if="item.type == 'quote'">
-              <p>{{ item.media.text }}</p>
+            <q-card-main v-if="item.type == 'quote' || item.type === 'outline' || item.type === 'idea' || item.type === 'illustration'">
+              <p v-if="item.type !== 'outline'">{{ item.media.text }}</p>
+              <ol v-if="item.type === 'outline' && item.media.numbered">
+                <li v-for="point in item.media.points" :key="point">
+                  {{ point.split('%%')[0] }}<br/><span class="q-caption">{{ point.split('%%')[1] }}</span>
+                </li>
+              </ol>
+              <ol v-if="item.type === 'outline' && !item.media.numbered">
+                <li v-for="point in item.media.points" :key="point">
+                  {{ point.split('%%')[0] }}<br/><span class="q-caption">{{ point.split('%%')[1] }}</span>
+                </li>
+              </ol>
               <p class="q-body-2"><span v-for="author in item.media.mediaid.author" :key="author">{{ author }} | </span>{{ item.media.mediaid.title }}</p>
             </q-card-main>
           </q-card>
@@ -46,7 +56,7 @@ export default {
     ResourcePreview
   },
   // name: 'ComponentName',
-  props: ['items', 'width', 'addModule'],
+  props: ['items', 'width', 'addModule', 'addButton'],
   data () {
     return {
       showItems: [],
@@ -54,18 +64,8 @@ export default {
       resourceOpen: false,
       resourceType: '',
       resource: '',
-      sizes: [
-        { columns: 1, gutter: 20 },
-        { mq: '800px', columns: 2, gutter: 20 },
-        { mq: '1000px', columns: 3, gutter: 20 },
-        { mq: '1200px', columns: 2, gutter: 20 }
-      ],
-      cardStyle: {
-        width: this.width
-      },
       selectedTypes: [
-        'book',
-        'quote'
+        'all'
       ],
       types: [
         {
@@ -129,25 +129,27 @@ export default {
   },
   watch: {
     'selectedTypes': function () {
-      this.showItems = this.items.filter(this.checkType)
+      if (this.selectedTypes[0] === 'all') {
+        this.showItems = this.items
+      } else {
+        this.showItems = this.items.filter(this.checkType)
+      }
     },
     'items': function (value) {
-      this.showItems = value
-      // this.pack()
-    },
-    'width': function () {
-      // this.pack()
+      if (this.selectedTypes[0] === 'all') {
+        this.showItems = value
+      } else {
+        this.showItems = value.filter(this.checkType)
+      }
     }
   },
-  // mounted () {
-  //   this.showItems = this.items.filter(this.checkType)
-  // },
-  updated () {
-    console.log('updated')
-    // console.log(this.$refs.bricks)
-    // this.pack()
+  mounted () {
+    this.init()
   },
   methods: {
+    init () {
+      this.showItems = this.items
+    },
     openItem (item, type) {
       console.log(item)
       console.log(type)
@@ -162,9 +164,6 @@ export default {
     closeResource () {
       this.resourceOpen = false
     },
-    // pack () {
-    //   this.$refs.bricks.pack()
-    // },
     checkType (item) {
       return this.selectedTypes.includes(item.type)
     }
@@ -187,12 +186,12 @@ export default {
   opacity: 1;
 }
 @media screen and (min-width: 800px) {
-  .bookl, .moviel, .imagel, .videol, .articlel, .notel, .quotel {
+  .media-cardl {
     width: 31%;
   }
 }
 @media screen and (min-width: 1200px) {
-  .bookl, .moviel, .imagel, .videol, .articlel, .notel, .quotel {
+  .media-cardl {
     width: 45%;
   }
   .resource-modal {